prevent access to non-root referenced fields
This works around the special case where we have replaced the doc
with the root doc. In this case any field reference that is not
'root based' should be ignored.
